Memory leak handled. There still a bug on main algoritm that must be found. Sometimes the ants find better solutions that possible.

+++ ant.h
@@ -2,6 +2,6 @@
 struct ant{
 	unsigned short int * planes_lt;
 	unsigned short int * planes_path;
-	unsigned int solution;
+	unsigned long long int solution;
 };
